refactor: user/statistics 패키지에 있던 사용자 풀이 이력 관련 클래스들을 user/history로 이동
