+++ src/main/java/com/upstage/devup/user/history/repository/UserSolvedHistoryRepository.java
@@ -1,4 +1,4 @@
-package com.upstage.devup.user.statistics.repository;
+package com.upstage.devup.user.history.repository;
 
 import com.upstage.devup.global.entity.UserAnswer;
 import org.springframework.data.domain.Page;
@@ -7,6 +7,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserAnswerHistoryRepository extends JpaRepository<UserAnswer, Long> {
+public interface UserSolvedHistoryRepository extends JpaRepository<UserAnswer, Long> {
+
     Page<UserAnswer> findByUserId(Long userId, Pageable pageable);
+
 }
