+++ src/test/java/com/upstage/devup/user/statistics/controller/UserStatControllerTest.java
@@ -3,7 +3,6 @@ package com.upstage.devup.user.statistics.controller;
 import com.upstage.devup.auth.config.AuthenticatedUser;
 import com.upstage.devup.auth.config.SecurityConfig;
 import com.upstage.devup.auth.config.jwt.JwtTokenProvider;
-import com.upstage.devup.user.statistics.dto.UserSolvedQuestionDto;
 import com.upstage.devup.user.statistics.dto.WrongNoteSummaryDto;
 import com.upstage.devup.user.statistics.service.UserAnswerStatService;
 import com.upstage.devup.user.statistics.service.UserWrongAnswerReadService;
@@ -48,48 +47,6 @@ class UserStatControllerTest {
 
     @MockitoBean
     private JwtTokenProvider jwtTokenProvider;
-
-    @Test
-    @DisplayName("사용자 문제 풀이 이력 조회 성공")
-    public void shouldGetSolvedQuestions() throws Exception {
-        // given
-        Long userId = 1L;
-
-        Integer pageNumber = 0;
-        int pageSize = 10;
-
-        UserSolvedQuestionDto dto = UserSolvedQuestionDto.builder()
-                .id(1L)
-                .questionId(1L)
-                .questionTitle("제목")
-                .category("카테고리")
-                .level("난이도")
-                .solvedAt(LocalDateTime.now())
-                .isCorrect(true)
-                .build();
-
-        List<UserSolvedQuestionDto> items = List.of(dto);
-        Pageable pageable = PageRequest.of(pageNumber, pageSize);
-        Page<UserSolvedQuestionDto> mockPage = new PageImpl<>(items, pageable, items.size());
-
-        when(userAnswerStatService.getUserSolvedQuestions(eq(userId), eq(pageNumber)))
-                .thenReturn(mockPage);
-
-        // when & then
-        mockMvc.perform(get("/api/stat/history")
-                        .param("pageNumber", String.valueOf(pageNumber))
-                        .with(getAuthentication(userId)))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.content").isArray())
-                .andExpect(jsonPath("$.content[0].id").value(dto.getId()))
-                .andExpect(jsonPath("$.content[0].questionId").value(dto.getQuestionId()))
-                .andExpect(jsonPath("$.content[0].questionTitle").value(dto.getQuestionTitle()))
-                .andExpect(jsonPath("$.content[0].category").value(dto.getCategory()))
-                .andExpect(jsonPath("$.content[0].level").value(dto.getLevel()))
-                .andExpect(jsonPath("$.content[0].firstSolvedAt").value(String.valueOf(dto.getSolvedAt())))
-                .andExpect(jsonPath("$.content[0].correct").value(String.valueOf(dto.isCorrect())))
-                .andExpect(jsonPath("$.totalElements").value(String.valueOf(items.size())));
-    }
 
     @Test
     @DisplayName("오답 노트 목록 조회 성공")
